backend/requirements.txt
<<<<<<< HEAD
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg[binary]
pydantic>=2.4.2
pydantic-settings>=2.0.3
bcrypt>=4.1.2
cffi>=1.16.0
python-jose[cryptography]==3.3.0
python-multipart>=0.0.6
reportlab>=4.0.7
Pillow>=10.2.0
python-dotenv>=1.0.1
=======
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
reportlab==4.0.7
>>>>>>> 01ae768219e574b7569fd6ef9d0968c847a4bb32
